velib-monitoring/requirements.txt
# === VÉLIB' MO# === WEB & DASHBOARD ===
streamlit>=1.28.0
plotly>=5.15.0
folium>=0.14.0
streamlit-folium>=0.13.0

# === UTILITIES === DÉPENDANCES PYTHON ===
# Compatible Python 3.8-3.13, macOS Homebrew, résout PEP 668

# === ESSENTIELS ===
requests>=2.28.0
python-dotenv>=1.0.0
pandas>=2.0.0
numpy>=1.24.0

# === STREAMING & MESSAGING ===
kafka-python>=2.0.0

# === DATABASE ===
pymongo>=4.0.0

# === WEB & DASHBOARD ===
streamlit>=1.28.0
plotly>=5.15.0
folium>=0.14.0
streamlit-folium>=0.13.0

# === UTILITIES ===
schedule>=1.2.0
pytz>=2023.3

# === DEVELOPMENT ===
pytest>=7.4.0
